requirements.txt
alpaca_trade_api
empyrical
kafka-python
python-dotenv
